+++ src/main/java/fr/iamacat/dangerzone_iamacatfr/entities/entity/ArcticScorpionInstance.java
@@ -683,6 +683,6 @@ public class ArcticScorpionInstance extends EntityMob {
     }
 
     public final void setAttacking(final int par1) {
-        this.dataWatcher.updateObject(20, (byte) par1);
+        this.dataWatcher.updateObject(20, par1);
     }
 }
