+++ src/main/java/fr/iamacat/dangerzone_iamacatfr/entities/entity/MooseInstance.java
@@ -409,7 +409,7 @@ public class MooseInstance extends EntityMob {
     }
 
     public void setAttacking(final int par1) {
-        this.dataWatcher.updateObject(20, (byte) par1);
+        this.dataWatcher.updateObject(20, par1);
     }
 
     @SideOnly(Side.CLIENT)
