+++ src/main/java/fr/iamacat/dangerzone_iamacatfr/entities/entity/TimberWolfInstance.java
@@ -442,7 +442,7 @@ public class TimberWolfInstance extends EntityMob {
     }
 
     public final void setAttacking(final int par1) {
-        this.dataWatcher.updateObject(20, (byte) par1);
+        this.dataWatcher.updateObject(20, par1);
     }
 
     public boolean getCanSpawnHere() {
